+++ Frontend/src/App.js
@@ -7,8 +7,8 @@ import Sginup from './pages/Signup'; //회원가입 페이지 추가
 import FindId from "./pages/FindId"; //아이디 찾기 페이지 추가
 import FindPwd from "./pages/FindPwd"; //비밀먼호 찾기 페이지 추가
 import Introduce from "./pages/Introduce"; //소개글 페이지 추가
-import MyPage from "./pages/Main/MyPage/MyPage";
-import MyPageEdit from "./pages/Main/MyPage/MyPageEdit";
+import MyPage from "./pages/MyPage/MyPage";
+import MyPageEdit from "./pages/MyPage/MyPageEdit";
 
 function App() {
   return (
