+++ Frontend/src/pages/Main/Main.css
@@ -101,7 +101,7 @@ p {
     right: 12px;
 }
 
-.hashtags {
+.popular_hashtags {
     display: flex;
     flex-direction: column;
     margin: 80px 0 120px 0;
@@ -109,14 +109,14 @@ p {
     width: 100%;
 }
 
-.hashtags_container {
+.popular_hashtags_container {
     display: flex;
     gap: 12px;
     margin-top: 20px;
     flex-wrap: wrap;
 }
 
-.hashtag {
+.popular_hashtag {
     font-family: 'PretendardBold';
     font-size: 18px;
     background-color: #D9D9D9;
@@ -125,7 +125,7 @@ p {
     padding: 21px 26px;
 }
 
-.hashtag:hover {
+.popular_hashtag:hover {
     background-color: #CEDEFF;
     cursor: pointer;
 }
